+++ package.json
@@ -14,7 +14,7 @@
     "lint:fix": "eslint --fix 'src/**/*.{js,json,ts}'",
     "format": "prettier --write 'src/**/*.{js,json,ts}' --config ./.prettierrc",
     "semantic-release": "semantic-release",
-    "test": "jest --coverage --watchAll",
+    "test": "jest --coverage",
     "test:ci": "jest --ci --passWithNoTests"
   },
   "husky": {
