--- src/constants.ts
@@ -1,12 +1,0 @@
-export const TRACKING_ID_KEY = 'fuul.tracking_id';
-export const PROJECT_ID_KEY = 'fuul.project_id';
-export const AFFILIATE_ID_KEY = 'fuul.affiliate_id';
-export const TRAFFIC_SOURCE_KEY = 'fuul.traffic_source';
-export const TRAFFIC_CATEGORY_KEY = 'fuul.traffic_category';
-export const TRAFFIC_TITLE_KEY = 'fuul.traffic_title';
-export const TRAFFIC_TAG_KEY = 'fuul.traffic_tag';
-export const TRAFFIC_REFERRER_URL = 'fuul.traffic_referrer_url';
-export const SENT_EVENT_ID_KEY = 'fuul.sent';
-export const SENT_EVENT_VALIDITY_PERIOD_MS = 60000;
-
-export const SEARCH_ENGINE_URLS = ['google.com', 'bing.com', 'yahoo.com'];
