--- src/types/index.ts
@@ -1,1 +1,0 @@
-export type { EventArgs, FuulSettings,UserMetadata } from './sdk';
