--- src/utils/queryParams.ts
@@ -1,9 +1,0 @@
-export const buildQueryParams = (args: Record<string, string>) => {
-  let queryParams = '';
-
-  Object.keys(args).forEach((key) => {
-    queryParams = queryParams === '' ? queryParams + `?${key}=${args[key]}` : queryParams + '&' + `${key}=${args[key]}`;
-  });
-
-  return queryParams;
-};
