+++ tsconfig.json
@@ -22,5 +22,5 @@
     },
     "typeRoots": ["node_modules/@types", "src/index.d.ts"]
   },
-  "include": ["src", "tests"]
+  "include": ["src"]
 }
